Initial React setup
This commit adds lots of new features into the repo. First thing is the package.json
and the project dependencies.
Webpack configuration was done alogside with babel and eslint setup. Project has now
two script build and serve, both coming from webpack configuration.
Src folder was created with basic files just to render something on the screen and
test if everything is working properly.
Jest was added to the project with a simple unit test on App.js.